requirements.txt
torch==2.7.1
librosa==0.9.2
espnet
espnet-tts-frontend==0.0.3
torchaudio==2.7.1
adjectiveanimalnumber==0.0.1
streamlit==1.46.1